refactor: WebsosoTheme 적용 및 기존 파일 제거
